Address comments and move cmd from patch file to cloud-init runcmd
