+++ engine/schema/pom.xml
@@ -122,8 +122,7 @@
                             <goal>wget</goal>
                         </goals>
                         <configuration>
-<!--                            <url>https://download.cloudstack.org/systemvm/${cs.version}/md5sum.txt</url>-->
-                            <url>http://10.0.3.122/systemvmtemplate/custom/cks-debian/${cs.version}/1/md5sum.txt</url>
+                            <url>https://download.cloudstack.org/systemvm/${cs.version}/md5sum.txt</url>
                             <outputDirectory>${basedir}/dist/systemvm-templates/</outputDirectory>
                             <skipCache>true</skipCache>
                             <overwrite>true</overwrite>
@@ -182,8 +181,7 @@
                                 </goals>
                                 <configuration>
                                     <checkSignature>true</checkSignature>
-<!--                                    <url>https://download.cloudstack.org/systemvm/${cs.version}/systemvmtemplate-${cs.version}.${patch.version}-kvm.qcow2.bz2</url>-->
-                                    <url>http://10.0.3.122/systemvmtemplate/custom/cks-debian/${cs.version}/1/systemvmtemplate-${cs.version}.${patch.version}-kvm.qcow2.bz2</url>
+                                    <url>https://download.cloudstack.org/systemvm/${cs.version}/systemvmtemplate-${cs.version}.${patch.version}-kvm.qcow2.bz2</url>
                                     <outputDirectory>${basedir}/dist/systemvm-templates/</outputDirectory>
                                     <md5>${kvm.checksum}</md5>
                                 </configuration>
@@ -195,8 +193,7 @@
                                 </goals>
                                 <configuration>
                                     <checkSignature>true</checkSignature>
-<!--                                    <url>https://download.cloudstack.org/systemvm/${cs.version}/systemvmtemplate-${cs.version}.${patch.version}-vmware.ova</url>-->
-                                    <url>http://10.0.3.122/systemvmtemplate/custom/cks-debian/${cs.version}/1/systemvmtemplate-${cs.version}.${patch.version}-vmware.ova</url>
+                                    <url>https://download.cloudstack.org/systemvm/${cs.version}/systemvmtemplate-${cs.version}.${patch.version}-vmware.ova</url>
                                     <outputDirectory>${basedir}/dist/systemvm-templates/</outputDirectory>
                                     <md5>${vmware.checksum}</md5>
                                 </configuration>
@@ -208,8 +205,7 @@
                                 </goals>
                                 <configuration>
                                     <checkSignature>true</checkSignature>
-<!--                                    <url>https://download.cloudstack.org/systemvm/${cs.version}/systemvmtemplate-${cs.version}.${patch.version}-xen.vhd.bz2</url>-->
-                                    <url>http://10.0.3.122/systemvmtemplate/custom/cks-debian/${cs.version}/1/systemvmtemplate-${cs.version}.${patch.version}-xen.vhd.bz2</url>
+                                    <url>https://download.cloudstack.org/systemvm/${cs.version}/systemvmtemplate-${cs.version}.${patch.version}-xen.vhd.bz2</url>
                                     <outputDirectory>${basedir}/dist/systemvm-templates/</outputDirectory>
                                     <md5>${xen.checksum}</md5>
                                 </configuration>
