+++ plugins/integrations/kubernetes-service/src/main/resources/conf/k8s-node.yml
@@ -260,5 +260,6 @@ write_files:
 runcmd:
   - chown -R cloud:cloud /home/cloud/.ssh
   - containerd config default > /etc/containerd/config.toml
+  - systemctl restart containerd
   - [ systemctl, start, setup-kube-system ]
   - [ systemctl, start, deploy-kube-system ]
